# examples/starter/backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: TinyCloud Starter API
  version: 0.1.0
  description: |
    Starter API for TinyCloud + OpenKey apps.
    Demonstrates JWT auth (OpenKey) + TinyCloud delegation patterns.

    **Two-layer auth:**
    1. Bearer JWT from OpenKey (OAuth 2.1 PKCE + passkeys)
    2. TinyCloud PortableDelegation granting scoped backend access
servers:
  - url: http://localhost:3001
    description: Local development

# ── Security ────────────────────────────────────────────────────────
security:
  - bearerAuth: []

# ── Paths ───────────────────────────────────────────────────────────
paths:
...
  /api/items:
    get:
      operationId: listItems
      summary: List items
      tags: [Items]
      parameters:
        - $ref: "#/components/parameters/StoreType"
        - name: limit
          in: query
          description: Maximum KV item values to fetch. Applies to the KV store only; defaults to 25 and is capped at 50. The installed TinyCloud SDK does not provide cursor pagination, so this limit is applied after listing matching keys.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 25
        - name: search
          in: query
          description: Search term for title and data (SQL store only)
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field (SQL store only)
          schema:
            type: string
            enum: [title, created_at, updated_at]
            default: created_at
        - name: dir
          in: query
          description: Sort direction (SQL store only)
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/NoDelegation"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/StoreError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

    post:
      operationId: createItem
      summary: Create a new item
      tags: [Items]
      parameters:
        - $ref: "#/components/parameters/StoreType"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: invalid_body
                message: "Request body must include a 'title' string field"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/NoDelegation"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/StoreError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Item ID

    get:
      operationId: getItem
      summary: Get a single item
      tags: [Items]
      parameters:
        - $ref: "#/components/parameters/StoreType"
      responses:
        "200":
          description: Item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/NoDelegation"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/StoreError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

    put:
      operationId: updateItem
      summary: Update an item
      tags: [Items]
      parameters:
        - $ref: "#/components/parameters/StoreType"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemInput"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: invalid_body
                message: "Request body must include at least 'title' or 'data'"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/NoDelegation"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/StoreError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

    delete:
      operationId: deleteItem
      summary: Delete an item
      tags: [Items]
      parameters:
        - $ref: "#/components/parameters/StoreType"
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "204":
          description: Item deleted (no content)
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/NoDelegation"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/StoreError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

# ── Components ──────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OpenKey JWT access token (OAuth 2.1 PKCE)
  parameters:
    StoreType:
      name: store
      in: query
      description: Storage backend to use
      schema:
        type: string
        enum: [kv, sql, duckdb]
        default: kv
    CsrfHeader:
      name: X-Requested-With
      in: header
      required: true
      description: CSRF protection header (required for state-changing methods)
      schema:
        type: string
        default: TinyBoilerplate
  schemas:
    Item:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        data:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateItemInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        data:
          type: string
    UpdateItemInput:
      type: object
      properties:
        title:
          type: string
        data:
          type: string
    ItemResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/Item"
    ItemListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        meta:
          $ref: "#/components/schemas/ItemListMeta"
    ItemListMeta:
      type: object
      description: KV list bounds and truncation metadata. Present for KV list responses. This describes a bounded sample from matching keys, not cursor pagination.
      required:
        [limit, returned, totalKeys, hasMore, truncated, limitedBy, supportsPagination, note]
      properties:
        limit:
          type: integer
          description: Effective server-side value-fetch limit after defaulting and max enforcement.
          example: 25
        returned:
          type: integer
          description: Number of item values returned after fetching and filtering.
          example: 25
        totalKeys:
          type: integer
          description: Number of keys matched by the KV prefix list.
          example: 42
        hasMore:
          type: boolean
          description: True when more KV keys matched than the server fetched. This does not imply a cursor is available.
          example: true
        truncated:
          type: boolean
          description: True when the response was capped by the effective limit.
          example: true
        limitedBy:
          type: string
          enum: [kv_value_fetch_limit]
          description: Stable reason the KV response is bounded.
          example: kv_value_fetch_limit
        supportsPagination:
          type: boolean
          const: false
          description: The installed TinyCloud SDK does not expose cursor/limit pagination for KV list.
          example: false
        note:
          type: string
          description: Human-readable note that this is a bounded sample, not true pagination.
          example: "KV list pagination is not available in the installed TinyCloud SDK; limit caps value fetches after listing matching keys."
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error description
  responses:
    Unauthenticated:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: missing_token
            message: Authorization header is required
...
    NoDelegation:
      description: No active delegation for this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: no_delegation
            message: Delegation required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: not_found
            message: "Item 'abc-123' not found"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: rate_limited
            message: Too many requests
...
    StoreError:
      description: Storage operation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: store_error
            message: Failed to list items
    GatewayTimeout:
      description: TinyCloud operation timed out
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: gateway_timeout
            message: TinyCloud operation timed out