# examples/starter/backend/src/routes/items.ts
import { Router } from "express";
import type { Request, Response } from "express";
import type { Item, CreateItemInput, UpdateItemInput, StoreType } from "@tinyboilerplate/core";
import type { DelegatedAccess } from "@tinyboilerplate/server";
import { ITEM_KV_PATH, ITEM_SQL_PATH, STARTER_APP_ID } from "../backend-policy.js";

const ITEM_KV_PREFIX = `${STARTER_APP_ID}/${ITEM_KV_PATH}`;
const ITEM_SQL_DATABASE = `${STARTER_APP_ID}/${ITEM_SQL_PATH}`;
const KV_LIST_FETCH_CONCURRENCY = 5;
const KV_LIST_DEFAULT_LIMIT = 25;
const KV_LIST_MAX_LIMIT = 50;
const KV_LIST_LIMITED_BY = "kv_value_fetch_limit";
const KV_LIST_PAGINATION_NOTE =
  "KV list pagination is not available in the installed TinyCloud SDK; limit caps value fetches after listing matching keys.";

// ── Items Router ─────────────────────────────────────────────────────

export function createItemsRouter() {
  const router = Router();

  // GET /api/items?store=kv|sql|duckdb
  router.get("/", async (req: Request, res: Response) => {
    const storeType = getStoreType(req);
    if (!req.delegatedAccess) {
      res.status(403).json({ error: "no_delegation", message: "Delegation required" });
      return;
    }
    const access = req.delegatedAccess;

    try {
      if (storeType === "sql" || storeType === "duckdb") {
        const svc = itemDatabase(access, storeType);
        await ensureTable(access, storeType);

        // SQL/DuckDB supports search + sort via query params
        const search = (req.query.search as string) ?? "";
        const sortBy = (req.query.sort as string) ?? "created_at";
        const sortDir = (req.query.dir as string) === "asc" ? "ASC" : "DESC";
        const validSort = ["title", "created_at", "updated_at"].includes(sortBy)
          ? sortBy
          : "created_at";

        let sql = "SELECT id, title, data, created_at, updated_at FROM items";
        const params: (string | number | null)[] = [];
        if (search) {
          sql += ` WHERE title LIKE ? OR data LIKE ?`;
          params.push(`%${search}%`, `%${search}%`);
        }
        sql += ` ORDER BY ${validSort} ${sortDir}`;

        const result = await svc.query(sql, params);
        if (!result.ok) throw new Error(`${storeType} query failed: ${result.error.message}`);
        const { rows, columns } = result.data;
        const items: Item[] = rows.map((row) => rowToItem(row, columns));
        res.json({ items });
      } else {
        // KV list returns { keys: string[] }, then get each value with bounded fanout
        const listResult = await access.kv.list({ prefix: ITEM_KV_PREFIX });
        if (!listResult.ok) {
          throw new Error(`KV list failed: ${listResult.error.message}`);
        }
        const keys = listResult.data.keys ?? [];
        const limit = getKVListLimit(req);
        const keysToFetch = keys.slice(0, limit);
        const results = await mapWithConcurrency(keysToFetch, KV_LIST_FETCH_CONCURRENCY, (key) =>
          access.kv.get(key),
        );
        const items: Item[] = results
          .filter(
            (r): r is typeof r & { ok: true; data: { data: unknown } } =>
              r.ok && r.data.data != null,
          )
          .map((r) => {
            const val = r.data.data;
            return (typeof val === "string" ? JSON.parse(val) : val) as Item;
          });
        res.json({
          items,
          meta: {
            limit,
            returned: items.length,
            totalKeys: keys.length,
            hasMore: keys.length > keysToFetch.length,
            truncated: keys.length > keysToFetch.length,
            limitedBy: KV_LIST_LIMITED_BY,
            supportsPagination: false,
            note: KV_LIST_PAGINATION_NOTE,
          },
        });
      }
    } catch (err) {
      handleStoreError(res, err, "list items");
    }
  });

  // POST /api/items?store=kv|sql|duckdb
  router.post("/", async (req: Request, res: Response) => {
    const storeType = getStoreType(req);
    if (!req.delegatedAccess) {
      res.status(403).json({ error: "no_delegation", message: "Delegation required" });
      return;
    }
    const access = req.delegatedAccess;
    const input: CreateItemInput = req.body;

    if (!input.title || typeof input.title !== "string") {
      res.status(400).json({
        error: "invalid_body",
        message: "Request body must include a 'title' string field",
      });
      return;
    }

    const now = new Date().toISOString();
    const item: Item = {
      id: crypto.randomUUID(),
      title: input.title,
      data: input.data,
      createdAt: now,
      updatedAt: now,
    };

    try {
      if (storeType === "sql" || storeType === "duckdb") {
        const svc = itemDatabase(access, storeType);
        await ensureTable(access, storeType);
        const insertResult = await svc.execute(
          `INSERT INTO items (id, title, data, created_at, updated_at) VALUES (?, ?, ?, ?, ?)`,
          [item.id, item.title, item.data ?? "", item.createdAt, item.updatedAt],
        );
        if (!insertResult.ok)
          throw new Error(`${storeType} insert failed: ${insertResult.error.message}`);
      } else {
        const putResult = await access.kv.put(itemKey(item.id), item);
        if (!putResult.ok) throw new Error(`KV put failed: ${putResult.error.message}`);
      }

      res.status(201).json({ item });
    } catch (err) {
      handleStoreError(res, err, "create item");
    }
  });

  // GET /api/items/:id?store=kv|sql|duckdb
  router.get("/:id", async (req: Request, res: Response) => {
    const storeType = getStoreType(req);
    if (!req.delegatedAccess) {
      res.status(403).json({ error: "no_delegation", message: "Delegation required" });
      return;
    }
    const access = req.delegatedAccess;
    const { id } = req.params;

    try {
      if (storeType === "sql" || storeType === "duckdb") {
        const svc = itemDatabase(access, storeType);
        await ensureTable(access, storeType);
        const result = await svc.query(
          `SELECT id, title, data, created_at, updated_at FROM items WHERE id = ?`,
          [id],
        );
        if (!result.ok) throw new Error(`${storeType} query failed: ${result.error.message}`);
        const { rows, columns } = result.data;
        if (rows.length === 0) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }
        res.json({ item: rowToItem(rows[0], columns) });
      } else {
        const result = await access.kv.get(itemKey(id));
        if (!result.ok || !result.data?.data) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }
        const raw = result.data.data;
        const item: Item = typeof raw === "string" ? JSON.parse(raw) : raw;
        res.json({ item });
      }
    } catch (err) {
      handleStoreError(res, err, "get item");
    }
  });

  // PUT /api/items/:id?store=kv|sql|duckdb
  router.put("/:id", async (req: Request, res: Response) => {
    const storeType = getStoreType(req);
    if (!req.delegatedAccess) {
      res.status(403).json({ error: "no_delegation", message: "Delegation required" });
      return;
    }
    const access = req.delegatedAccess;
    const { id } = req.params;
    const input: UpdateItemInput = req.body;

    if (!input.title && !input.data && input.data !== "") {
      res.status(400).json({
        error: "invalid_body",
        message: "Request body must include at least 'title' or 'data'",
      });
      return;
    }

    try {
      const now = new Date().toISOString();

      if (storeType === "sql" || storeType === "duckdb") {
        const svc = itemDatabase(access, storeType);
        await ensureTable(access, storeType);

        // Check if exists
        const existing = await svc.query(`SELECT id FROM items WHERE id = ?`, [id]);
        if (!existing.ok || existing.data.rows.length === 0) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }

        const setClauses: string[] = [];
        const updateParams: (string | number | null)[] = [];
        if (input.title !== undefined) {
          setClauses.push("title = ?");
          updateParams.push(input.title);
        }
        if (input.data !== undefined) {
          setClauses.push("data = ?");
          updateParams.push(input.data);
        }
        setClauses.push("updated_at = ?");
        updateParams.push(now);
        updateParams.push(id);

        const updateResult = await svc.execute(
          `UPDATE items SET ${setClauses.join(", ")} WHERE id = ?`,
          updateParams,
        );
        if (!updateResult.ok)
          throw new Error(`${storeType} update failed: ${updateResult.error.message}`);

        // Fetch the updated item
        const result = await svc.query(
          `SELECT id, title, data, created_at, updated_at FROM items WHERE id = ?`,
          [id],
        );
        if (!result.ok) throw new Error(`${storeType} query failed: ${result.error.message}`);
        const { rows, columns } = result.data;
        if (rows.length === 0) throw new Error("Item disappeared after update");
        res.json({ item: rowToItem(rows[0], columns) });
      } else {
        // KV: read-modify-write
        const result = await access.kv.get(itemKey(id));
        if (!result.ok || !result.data?.data) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }

        const raw = result.data.data;
        const existing: Item = typeof raw === "string" ? JSON.parse(raw) : raw;
        const updated: Item = {
          ...existing,
          title: input.title ?? existing.title,
          data: input.data !== undefined ? input.data : existing.data,
          updatedAt: now,
        };

        const putResult = await access.kv.put(itemKey(id), updated);
        if (!putResult.ok) throw new Error(`KV put failed: ${putResult.error.message}`);
        res.json({ item: updated });
      }
    } catch (err) {
      handleStoreError(res, err, "update item");
    }
  });

  // DELETE /api/items/:id?store=kv|sql|duckdb
  router.delete("/:id", async (req: Request, res: Response) => {
    const storeType = getStoreType(req);
    if (!req.delegatedAccess) {
      res.status(403).json({ error: "no_delegation", message: "Delegation required" });
      return;
    }
    const access = req.delegatedAccess;
    const { id } = req.params;

    try {
      if (storeType === "sql" || storeType === "duckdb") {
        const svc = itemDatabase(access, storeType);
        await ensureTable(access, storeType);

        // Check if exists
        const existing = await svc.query(`SELECT id FROM items WHERE id = ?`, [id]);
        if (!existing.ok || existing.data.rows.length === 0) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }

        const deleteResult = await svc.execute(`DELETE FROM items WHERE id = ?`, [id]);
        if (!deleteResult.ok)
          throw new Error(`${storeType} delete failed: ${deleteResult.error.message}`);
      } else {
        // KV: check existence, then delete
        const result = await access.kv.get(itemKey(id));
        if (!result.ok || !result.data?.data) {
          res.status(404).json({
            error: "not_found",
            message: `Item '${id}' not found`,
          });
          return;
        }

        await access.kv.delete(itemKey(id));
      }

      res.status(204).send();
    } catch (err) {
      handleStoreError(res, err, "delete item");
    }
  });

  return router;
}

// ── Helpers ──────────────────────────────────────────────────────────

function getStoreType(req: Request): StoreType {
  const store = req.query.store as string | undefined;
  if (store === "sql") return "sql";
  if (store === "duckdb") return "duckdb";
  return "kv";
}

function itemKey(id: string): string {
  return `${ITEM_KV_PREFIX}${id}`;
}

function getKVListLimit(req: Request): number {
  const value = Array.isArray(req.query.limit) ? req.query.limit[0] : req.query.limit;
  const requested = typeof value === "string" ? Number(value) : Number.NaN;
  if (!Number.isFinite(requested) || requested < 1) return KV_LIST_DEFAULT_LIMIT;
  return Math.min(Math.floor(requested), KV_LIST_MAX_LIMIT);
}

type DatabaseService = DelegatedAccess["sql"] | DelegatedAccess["duckdb"];

function itemDatabase(access: DelegatedAccess, storeType: "sql" | "duckdb"): DatabaseService {
  const service = storeType === "duckdb" ? access.duckdb : access.sql;
  const maybeScoped = service as DatabaseService & { db?: (name: string) => DatabaseService };
  return typeof maybeScoped.db === "function" ? maybeScoped.db(ITEM_SQL_DATABASE) : service;
}

/**
 * Ensure the items table exists. Runs CREATE TABLE IF NOT EXISTS
 * at most once per DelegatedAccess + store type combination.
 */
const tableCreated = new WeakMap<object, Set<string>>();

async function ensureTable(access: DelegatedAccess, storeType: "sql" | "duckdb"): Promise<void> {
  let created = tableCreated.get(access);
  if (!created) {
    created = new Set();
    tableCreated.set(access, created);
  }
  if (created.has(storeType)) return;

  const svc = itemDatabase(access, storeType);
  const result = await svc.execute(`
    CREATE TABLE IF NOT EXISTS items (
      id TEXT PRIMARY KEY,
      title TEXT NOT NULL,
      data TEXT DEFAULT '',
      created_at TEXT NOT NULL,
      updated_at TEXT NOT NULL
    )
  `);
  if (!result.ok) {
    throw new Error(`Failed to create items table (${storeType}): ${result.error.message}`);
  }

  created.add(storeType);
}

/**
 * Convert a SQL row to an Item.
 * QueryResponse rows are arrays (not objects), so we map by column index.
 */
function rowToItem(row: unknown[], columns: string[]): Item {
  const obj: Record<string, unknown> = {};
  columns.forEach((col, i) => {
    obj[col] = row[i];
  });
  return {
    id: obj.id as string,
    title: obj.title as string,
    data: (obj.data as string) ?? undefined,
    createdAt: obj.created_at as string,
    updatedAt: obj.updated_at as string,
  };
}

async function mapWithConcurrency<T, R>(
  values: T[],
  concurrency: number,
  mapper: (value: T) => Promise<R>,
): Promise<R[]> {
  const results = new Array<R>(values.length);
  let nextIndex = 0;
  const workerCount = Math.min(Math.max(1, concurrency), values.length);

  async function worker(): Promise<void> {
    while (nextIndex < values.length) {
      const index = nextIndex;
      nextIndex += 1;
      results[index] = await mapper(values[index]);
    }
  }

  await Promise.all(Array.from({ length: workerCount }, worker));
  return results;
}

function handleStoreError(res: Response, err: unknown, operation: string): void {
  console.error(`[items] ${operation} failed:`, err);
  const isTimeout = err instanceof Error && err.message.includes("timed out");
  res.status(isTimeout ? 504 : 500).json({
    error: isTimeout ? "gateway_timeout" : "store_error",
    message: isTimeout ? "TinyCloud operation timed out" : `Failed to ${operation}`,
  });
}